+++ example/DirectoryExample.php
@@ -6,38 +6,11 @@
  */
 
 require __DIR__.'/../vendor/autoload.php';
+require __DIR__.'/lib/example-helper.php';
 
 use LDL\File\Directory;
-use LDL\File\Helper\DirectoryHelper;
-use LDL\File\Helper\PathHelper;
-use LDL\File\File;
 
-$path = PathHelper::createAbsolutePath(sys_get_temp_dir(), 'ldl-file-common-test');
-
-/**
- * In case execution was aborted and the script failed to delete the directory
- */
-if(is_dir($path)){
-    DirectoryHelper::delete($path);
-}
-
-$tempDir = Directory::create($path);
-
-echo "Create some random files and directories in the system's temp dir and append them to a Directory collection";
-
-$directories = array_map(static function($file) use ($tempDir){
-    /**
-     * Randomize whether to create a file or a directory
-     */
-    $file = random_int(0,1) === 1 ? $tempDir->mkdir((string) $file) : File::create("{$tempDir}/{$file}.txt", uniqid('', true));
-
-    /**
-     * If a directory was created, create a nested directory, we do this to test recursive deletion
-     */
-    if($file instanceof Directory){
-        $file->mkdir((string)random_int(100,200));
-    }
-}, range(1,20));
+$tempDir = createTestFiles();
 
 echo "Get directory tree and print it out\n";
 echo "####################################################\n";