+++ src/File/Collection/DirectoryCollection.php
@@ -4,10 +4,12 @@ namespace LDL\File\Collection;
 
 use LDL\File\Collection\Contracts\DirectoryCollectionInterface;
 use LDL\File\Collection\Traits\AppendDirectoryAsStringTrait;
+use LDL\File\Contracts\DirectoryInterface;
 use LDL\File\Directory;
 use LDL\Type\Collection\AbstractTypedCollection;
 use LDL\Type\Collection\Traits\Validator\AppendValueValidatorChainTrait;
 use LDL\Validators\ClassComplianceValidator;
+use LDL\Validators\InterfaceComplianceValidator;
 
 final class DirectoryCollection extends AbstractTypedCollection implements DirectoryCollectionInterface
 {
@@ -20,7 +22,7 @@ final class DirectoryCollection extends AbstractTypedCollection implements Direc
 
         $this->getAppendValueValidatorChain()
             ->getChainItems()
-            ->append(new ClassComplianceValidator(Directory::class,true))
+            ->append(new InterfaceComplianceValidator(DirectoryInterface::class))
             ->lock();
     }