+++ src/File/Collection/JsonFileCollection.php
@@ -4,12 +4,12 @@ namespace LDL\File\Collection;
 
 use LDL\File\Collection\Contracts\FileCollectionInterface;
 use LDL\File\Collection\Traits\AppendFileAsStringTrait;
-use LDL\File\File;
+use LDL\File\Contracts\FileInterface;
 use LDL\File\Validator\JsonFileValidator;
 use LDL\File\Validator\ReadableFileValidator;
 use LDL\Type\Collection\AbstractTypedCollection;
 use LDL\Type\Collection\Traits\Validator\AppendValueValidatorChainTrait;
-use LDL\Validators\ClassComplianceValidator;
+use LDL\Validators\InterfaceComplianceValidator;
 
 final class JsonFileCollection extends AbstractTypedCollection implements FileCollectionInterface
 {
@@ -21,7 +21,7 @@ final class JsonFileCollection extends AbstractTypedCollection implements FileCo
         $this->getAppendValueValidatorChain()
             ->getChainItems()
             ->appendMany([
-                new ClassComplianceValidator(File::class, true),
+                new InterfaceComplianceValidator(FileInterface::class),
                 new ReadableFileValidator(),
                 new JsonFileValidator()
             ])