+++ src/File/Collection/ReadWriteFileCollection.php
@@ -3,6 +3,7 @@
 namespace LDL\File\Collection;
 
 use LDL\File\Collection\Contracts\FileCollectionInterface;
+use LDL\File\Contracts\FileInterface;
 use LDL\File\File;
 use LDL\File\Validator\FileExistsValidator;
 use LDL\File\Validator\ReadableFileValidator;
@@ -10,7 +11,7 @@ use LDL\File\Validator\WritableFileValidator;
 use LDL\Type\Collection\AbstractTypedCollection;
 use LDL\Type\Collection\Traits\Validator\AppendValueValidatorChainTrait;
 use LDL\File\Collection\Traits\AppendFileAsStringTrait;
-use LDL\Validators\ClassComplianceValidator;
+use LDL\Validators\InterfaceComplianceValidator;
 
 final class ReadWriteFileCollection extends AbstractTypedCollection implements FileCollectionInterface
 {
@@ -22,7 +23,7 @@ final class ReadWriteFileCollection extends AbstractTypedCollection implements F
         $this->getAppendValueValidatorChain()
             ->getChainItems()
             ->appendMany([
-                new ClassComplianceValidator(File::class,true),
+                new InterfaceComplianceValidator(FileInterface::class),
                 new FileExistsValidator(),
                 new ReadableFileValidator(),
                 new WritableFileValidator()