+++ src/File/Collection/WritableFileCollection.php
@@ -4,12 +4,12 @@ namespace LDL\File\Collection;
 
 use LDL\File\Collection\Contracts\FileCollectionInterface;
 use LDL\File\Collection\Traits\AppendFileAsStringTrait;
-use LDL\File\File;
+use LDL\File\Contracts\FileInterface;
 use LDL\File\Validator\FileExistsValidator;
 use LDL\File\Validator\WritableFileValidator;
 use LDL\Type\Collection\AbstractTypedCollection;
 use LDL\Type\Collection\Traits\Validator\AppendValueValidatorChainTrait;
-use LDL\Validators\ClassComplianceValidator;
+use LDL\Validators\InterfaceComplianceValidator;
 
 final class WritableFileCollection extends AbstractTypedCollection implements FileCollectionInterface
 {
@@ -21,7 +21,7 @@ final class WritableFileCollection extends AbstractTypedCollection implements Fi
         $this->getAppendValueValidatorChain()
             ->getChainItems()
             ->appendMany([
-                new ClassComplianceValidator(File::class, true),
+                new InterfaceComplianceValidator(FileInterface::class),
                 new FileExistsValidator(),
                 new WritableFileValidator()
             ])